+++ R/interpret_bf.R
@@ -1,7 +1,10 @@
+#' Bayes Factor Interpretation
+#'
 #' Return the interpretation of a Bayes Factor.
 #'
 #' @param x Bayes Factor.
-#' @param label_only Keep only the size classification.
+#' @param direction Include direction (against / in favour).
+#' @param bf Include Bayes Factor.
 #'
 #' @return The interpretation according to Jeffreys (1961).
 #'
@@ -17,13 +20,15 @@
 #'  \item{Jarosz, A. F., & Wiley, J. (2014). What are the odds? A practical guide to computing and reporting Bayes factors. The Journal of Problem Solving, 7(1), 2.}
 #'  }
 #' @export
-interpret_bf <- function(x, label_only=FALSE) {
+interpret_bf <- function(x, direction=TRUE, bf=TRUE) {
   if (x < 1) {
     x <- 1 / abs(x)
-    direction <- "against"
+    dir <- "against"
   } else {
-    direction <- "in favour of"
+    dir <- "in favour of"
   }
+
+
 
   interpretation <- ifelse(abs(x) > 100, "extreme evidence",
     ifelse(abs(x) >= 30, "very strong evidence",
@@ -35,8 +40,36 @@ interpret_bf <- function(x, label_only=FALSE) {
     )
   )
 
-  if (label_only == FALSE) {
-    interpretation <- paste(interpretation, direction)
+  if(bf == TRUE){
+    bf <- paste0("(", format_bf(x), ")")
+    interpretation <- paste(interpretation, bf)
   }
+  if (direction == TRUE) {
+    interpretation <- paste(interpretation, dir)
+  }
+
   return(interpretation)
 }
+
+
+
+
+
+#' Bayes factor formatting
+#'
+#' Bayes factor formatting
+#'
+#' @param bf Bayes Factor.
+#' @param max Treshold for maximum.
+#'
+#' @export
+format_bf <- function(bf, max=100){
+
+  if(bf > max){
+    bf <- paste0("BF > ", max)
+  } else{
+    bf <- paste0("BF = ", format_digit(bf))
+  }
+  return(bf)
+}
+